refactor by moving to package/deps
